+++ src/main/java/vendingmachine/Model/Validator/ProductValidator.java
@@ -1,4 +1,4 @@
-package vendingmachine.Model;
+package vendingmachine.Model.Validator;
 
 import java.util.Arrays;
